feat(#297): Phase 4 (partial) - Move inspection components to shared
- Created components directory structure under src/ui/diagnostics/components/
- Moved all inspection components from src/ui/inspection/ to src/ui/diagnostics/components/shared/
- Updated imports in DiagnosticsPanel.tsx to use new shared location
- Updated test imports in inspection-components.spec.tsx and inspection-panel.spec.tsx
- All 102 tests passing

Phase 4 in progress. Next: Extract StatsOverview, ContentPanels, LogsPanel, FooterPanel components.
