+++ src/ui/diagnostics/components/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Diagnostics Components
+ * 
+ * Barrel export for all diagnostics UI components.
+ * These components are extracted from DiagnosticsPanel.tsx for better modularity.
+ * 
+ * @see docs/refactoring/diagnostics-modularity-strategy.md
+ * @see docs/refactoring/diagnostics-before-after.md
+ */
+
+// Will be populated as components are extracted
+
