+++ src/ui/diagnostics/components/shared/CodeBlock.tsx
