+++ src/ui/diagnostics/components/shared/CollapsibleSection.tsx
