+++ src/ui/diagnostics/components/shared/InspectionPanel.tsx
