+++ src/ui/diagnostics/components/shared/JSONViewer.tsx
