+++ src/ui/diagnostics/components/shared/PluginInfoInspector.tsx
