+++ src/ui/diagnostics/components/shared/index.ts
