+++ src/ui/diagnostics/components/shared/inspection.css
